Made lenspop a package; re-organized imageSim package while trying to understand it; checked stellarpop can be imported. Added requirements.txt and notes on setup in README
