--- __init__.py
@@ -1,4 +1,0 @@
-from PopulationFunctions import *
-from MakeLensPop import *
-from Surveys import *
-from FastLensSim import *
