+++ imageSim/SBmodels.py
@@ -1,6 +1,7 @@
-import profiles
-from pointSource import PixelizedModel as PM, GaussianModel as GM
 from math import pi
+
+from .profiles import _Sersic, _Gauss
+from .pointSource import PixelizedModel as PM, GaussianModel as GM
 
 def cnts2mag(cnts,zp):
     from math import log10
@@ -54,8 +55,8 @@ class SBModel:
             self.__setattr__(key,self.vmap[key].value)
 
 
-class Sersic(SBModel,SBProfiles.Sersic):
-    _baseProfile = SBProfiles.Sersic
+class Sersic(SBModel, _Sersic):
+    _baseProfile = _Sersic
     _SBkeys = [['amp','n','pa','q','re','x','y'],
                 ['logamp','n','pa','q','re','x','y'],
                 ['amp','n','q','re','theta','x','y'],
@@ -63,10 +64,6 @@ class Sersic(SBModel,SBProfiles.Sersic):
 
     def __init__(self,name,pars,convolve=0):
         SBModel.__init__(name,pars,convolve)
-
-        else:
-            self.__dict__[key] = value
-
 
     def getMag(self,amp,zp):
         from scipy.special import gamma
@@ -81,7 +78,7 @@ class Sersic(SBModel,SBProfiles.Sersic):
         return self.getMag(self.amp,zp)
 
 
-class Gauss(profiles._Gauss):
+class Gauss(_Gauss):
 
     def __init__(self,name,var=None,const=None,convolve=0):
         if const is None:
@@ -98,7 +95,7 @@ class Gauss(profiles._Gauss):
             import sys
             print "Not all parameters defined!"
             sys.exit()
-        profiles._Gauss.__init__(self)
+        imageSim._Gauss.__init__(self)
         self.invar = var
         self.keys = keys
         self.values = {'r0':None}
@@ -150,7 +147,7 @@ class Gauss(profiles._Gauss):
 
 
 
-class PointSource(GM,PM):
+class PointSource(GM, PM):
     def __init__(self,name,model,var=None,const=None):
         if const is None:
             const = {}
@@ -177,7 +174,7 @@ class PointSource(GM,PM):
             self.ispix = True
         self.setValues()
         self.name = name
-        self.convolve = None 
+        self.convolve = None
 
     def __setattr__(self,key,value):
         if key=='logamp':
@@ -210,4 +207,3 @@ class PointSource(GM,PM):
         for key in self.vmap:
             self.values[self.vmap[key]] = pars[key]
         self.setValues()
-
